chore: move types to the connector
