+++ src/types/index.ts
@@ -3,4 +3,3 @@ export * from './connector';
 export * from './widget';
 export * from './insights';
 export * from './algoliasearch';
-export * from './infiniteHits';
